+++ src/components/Homepage/events/elements.js
@@ -1,0 +1,31 @@
+import styled from 'styled-components'
+import remcalc from 'remcalc'
+import breakpoint from 'styled-components-breakpoint'
+import { Col } from '../../grid'
+import { BodyPrimary } from '../../Typography'
+
+export const EventsColumn = styled(Col)`
+  padding-left: 0;
+`
+
+export const EventWrapper = styled.header`
+  padding: ${remcalc(18)} ${remcalc(24)} 0;
+
+  ${breakpoint('tablet')`
+    padding: ${remcalc(24)} ${remcalc(36)} 0;
+  `} ${breakpoint('desktop')`
+    padding-top: ${remcalc(24)};
+    padding-left: ${remcalc(36)};
+  `};
+`
+
+export const FixedWidthBodyPrimary = styled(BodyPrimary)`
+  max-width: ${remcalc(380)};
+`
+
+export const FeaturedEventWrapper = styled.section`
+  background-color: #${props => props.color};
+  ${breakpoint('smallTablet')`
+    margin-top: ${remcalc(175)}
+  `}
+`
