+++ src/pages/index.js
@@ -7,7 +7,7 @@ import Layout from '../components/layout'
 import SEOText from '../components/Homepage/seoText'
 import Companies from '../components/Homepage/companies'
 import Blog from '../components/Homepage/blog'
-import Events from '../components/Homepage/events'
+import Events from '../components/Homepage/events/index'
 import Jobs from '../components/Homepage/jobs'
 import CaseStudyPreview from '../components/Common/CaseStudyCards/CaseStudyPreview'
 import Services from '../components/Homepage/services'
